Inline SQLite command release helper into build script and drop dead file
